Create onnx_disparity SQL Check
This is a set of SQL Statements that provides a similarity index ONNX_DISPARITY between two image files, when the original ONNX Training Models are unavailable for comparison. Can be used for the same image generated from different training sessions. 

Table IMAGE1 contains the image BLOB and available Vector for an Image.

Table IMAGE2 contains the image BLOB and available Vector for an Image (maybe same as IMAGE1).

User Enters Directory of Image files.

ONNX_DISPARITY senses the similarity of Training Models used for IMAGE1 and IMAGE2.